+++ noodles-gff/src/line_buf.rs
@@ -17,16 +17,6 @@ pub enum LineBuf {
     Record(RecordBuf),
 }
 
-impl fmt::Display for LineBuf {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Directive(directive) => write!(f, "{directive}"),
-            Self::Comment(comment) => write!(f, "{COMMENT_PREFIX}{comment}"),
-            Self::Record(record) => write!(f, "{record}"),
-        }
-    }
-}
-
 /// An error returns when a raw GFF line fails to parse.
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub enum ParseError {
@@ -75,18 +65,6 @@ impl FromStr for LineBuf {
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn test_fmt() {
-        let line = LineBuf::Directive(DirectiveBuf::GffVersion(Default::default()));
-        assert_eq!(line.to_string(), "##gff-version 3");
-
-        let line = LineBuf::Comment(String::from("format: gff3"));
-        assert_eq!(line.to_string(), "#format: gff3");
-
-        let line = LineBuf::Record(RecordBuf::default());
-        assert_eq!(line.to_string(), ".\t.\t.\t1\t1\t.\t.\t.\t.")
-    }
 
     #[test]
     fn test_from_str() {